requirements.txt
fastapi==0.120.3
uvicorn==0.33.0
python-multipart==0.0.20
numpy==1.24.4
psycopg2-binary==2.9.10
sqlalchemy==2.0.44
pydantic==2.10.6
requests==2.32.4
streamlit==1.40.1
matplotlib==3.7.5
python-dotenv